requirements.txt
altair==5.4.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
filelock==3.16.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.7.0
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
protobuf==5.28.1
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.12.2
sortedcontainers==2.4.0
stack-data==0.6.3
streamlit==1.38.0
tenacity==8.5.0
textdistance==4.6.3
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
